+++ packages/fl_select/lib/src/select/wrap_select.dart
@@ -206,7 +206,7 @@ class WrapSelectSkeleton extends StatelessWidget {
         Flexible(
           child: Padding(
             padding: const EdgeInsets.symmetric(vertical: 10, horizontal: 12),
-            child: SelectChipBarSkeleton(
+            child: SelectWrapViewSkeleton(
               itemCount: itemCount,
               padding: padding,
               spacing: spacing,